Refactor neuromodulation
