+++ hypothesis/nn/neuromodulation/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseNeuromodulatedModule
+from .elu import NeuromodulatedELU
+from .relu import NeuromodulatedReLU
+from .selu import NeuromodulatedSELU
+from .tanh import NeuromodulatedELU
